requirements.txt
pinecone-client==4.1.2
langchain-pinecone==0.1.1
langchain-community==0.2.6
langchain-huggingface==0.0.3
unstructured==0.14.9
accelerate==0.32.1
pypdf